+++ src/main/java/com/escola/projeto/domain/entities/Nota.java
@@ -10,6 +10,7 @@ import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.DBRef;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -23,7 +24,7 @@ public class Nota {
     private Estudante estudante;
     @DBRef
     private Atividade atividade;
-    private Double nota;
+    private BigDecimal nota;
 
     public Nota(NotaRequestDTO notaRequestDTO) {
         this.setId(notaRequestDTO.getId());
